requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
pytest==6.2.5
pytest-cov==2.12.1
requests==2.26.0
python-multipart==0.0.5
email-validator==1.1.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==0.19.0
httpx==0.19.0
# Appium dependencies
Appium-Python-Client==2.6.0
selenium==4.1.0
pytest-xdist==2.4.0
pytest-html==3.1.1 